Create README.md
This folder contains my work for the Credit Card Fraud Detection competition on Kaggle. The competition involves building a model to identify fraudulent credit card transactions. Included is the code used to create the model and create a submission. Both the training and testing csv exceeded the file size limit and are not included.